fix: Change lastRequestTime type from int to float
Fixes PHP deprecation warning about implicit float to int conversion

+++ app/Services/ViewDnsService.php
@@ -7,7 +7,7 @@ namespace App\Services;
 class ViewDnsService
 {
     private static ?string $apiKey = null;
-    private static int $lastRequestTime = 0;
+    private static float $lastRequestTime = 0;
     private static int $rateLimit = 10; // 默认每秒10次
 
     /**
